# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for FindMe application.",
        "title": "FindMe API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/user/connect-github": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "An redirecting endpoint for connecting a github account to the current user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Connecting github account to user account",
                "responses": {
                    "307": {
                        "description": "Redirecting",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    }
                }
            }
        },
...
        "/api/user/view-repo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "An endpoint for viewing the user public repo to tag to a project.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Git"
                ],
                "summary": "View Github public repo for current user",
                "responses": {
                    "200": {
                        "description": "Repo fetched",
                        "schema": {
                            "$ref": "#/definitions/schema.DocViewRepos"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "422": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/auth/github/callback": {
            "get": {
                "description": "An endpoint for selecting the callback for login, connect github endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Callback for the github sign-in / connect",
                "responses": {
                    "200": {
                        "description": "Token generated",
                        "schema": {
                            "$ref": "#/definitions/schema.DocTokenResponse"
                        }
                    },
                    "202": {
                        "description": "Github Connected",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid id",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "409": {
                        "description": "Existing record",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    }
                }
            }
        },
...
        "/github-signup": {
            "get": {
                "description": "A redirecting endpoint for sign-in / sign-up with github",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Signing up user using github",
                "responses": {
                    "307": {
                        "description": "Redirecting",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/schema.DocNormalResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.DocNormalResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "schema.DocTokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "schema.DocViewRepos": {
            "type": "object",
            "properties": {
                "repos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.ViewRepo"
                    }
                }
            }
        },
...
        "schema.ViewRepo": {
            "type": "object",
            "properties": {
                "html_url": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}